feat(examples): port input to client input
