+++ guest/rust/examples/basics/input/Cargo.toml
@@ -10,9 +10,9 @@ version = {workspace = true}
 ambient_api = {workspace = true}
 
 [[bin]]
-name = "input_server"
-path = "src/server.rs"
-required-features = ["server"]
+name = "input_client"
+path = "src/client.rs"
+required-features = ["client"]
 
 [features]
 client = ["ambient_api/client"]
